src/requirements.txt
pytesseract
Pillow
opencv-python-headless
opencv-contrib-python-headless
numpy